+++ tests/Probabilities/BinomialDistributionTest.php
@@ -6,7 +6,6 @@ namespace Tests\Innmind\Math\Probabilities;
 use Innmind\Math\{
     Probabilities\BinomialDistribution,
     Algebra\Number,
-    Algebra\Integer,
     Algebra\Real,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
@@ -17,7 +16,7 @@ class BinomialDistributionTest extends TestCase
     {
         $law = BinomialDistribution::of(Real::of(0.5));
 
-        $probability = $law(Integer::of(9), Integer::of(2));
+        $probability = $law(9, 2);
 
         $this->assertInstanceOf(Number::class, $probability);
         $this->assertSame(0.0703125, $probability->value());