+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Requires PHP `8.4`
 - `Innmind\Math\DefinitionSet\Set` is now a final class, all previous implementations are now flagged as internal
 - `Innmind\Math\Monoid\*` are now enums
+- `Innmind\Math\Probabilities\BinomialDistribution::__invoke()` arguments now expects `int`s
 
 ### Fixed